Ref: Created /message_text and moved files inside
