+++ app/src/components/market/sections/common_sections/message_text/market_resolution_message/img/Won.tsx
