+++ app/src/components/market/sections/common_sections/message_text/token_item/index.tsx
