+++ app/src/components/market/sections/common_sections/message_text/warning_message/index.tsx
