+++ app/src/components/market/sections/market_buy/market_buy.tsx
@@ -23,9 +23,9 @@ import { OutcomeTable } from '../../common/outcome_table'
 import { TransactionDetailsCard } from '../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../common/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
-import { WarningMessage } from '../../common/warning_message'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
+import { WarningMessage } from '../common_sections/message_text/warning_message'
 
 const WarningMessageStyled = styled(WarningMessage)`
   margin-top: 20px;
