+++ app/src/components/market/sections/market_create/steps/items/create_scalar_market.tsx
@@ -12,8 +12,8 @@ import { DateField, FormRow, Textfield } from '../../../../../common'
 import { BigNumberInput, BigNumberInputReturn } from '../../../../../common/form/big_number_input'
 import { QuestionInput } from '../../../../../common/form/question_input'
 import { Categories } from '../../../../common/categories'
-import { WarningMessage } from '../../../../common/warning_message'
 import { Arbitrators } from '../../../common_sections/arbitrators/arbitrators'
+import { WarningMessage } from '../../../common_sections/message_text/warning_message'
 
 import { ButtonCategory, ButtonCategoryTextOverflow } from './ask_question_step'
