+++ app/src/components/market/sections/market_profile/market_status/closed.tsx
@@ -27,11 +27,11 @@ import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
 import { ModalTransactionWrapper } from '../../../../modal'
 import { MarginsButton } from '../../../common/common_styled'
-import MarketResolutionMessage from '../../../common/market_resolution_message'
 import { OutcomeTable } from '../../../common/outcome_table'
 import { ViewCard } from '../../../common/view_card'
 import { MarketScale } from '../../common_sections/card_bottom_details/market_scale'
 import { MarketTopDetailsClosed } from '../../common_sections/card_top_details/market_top_details_closed'
+import MarketResolutionMessage from '../../common_sections/message_text/market_resolution_message'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
 import { MarketNavigation } from '../../market_navigation'
