+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -16,9 +16,9 @@ import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
 import { OutcomeTable } from '../../../common/outcome_table'
 import { ViewCard } from '../../../common/view_card'
-import { WarningMessage } from '../../../common/warning_message'
 import { MarketScale } from '../../common_sections/card_bottom_details/market_scale'
 import { MarketTopDetailsOpen } from '../../common_sections/card_top_details/market_top_details_open'
+import { WarningMessage } from '../../common_sections/message_text/warning_message'
 import { MarketBondContainer } from '../../market_bond/market_bond_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
 import { MarketNavigation } from '../../market_navigation'
