+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -18,8 +18,8 @@ import { TokenBalance } from '../../common/token_balance'
 import { TransactionDetailsCard } from '../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../common/transaction_details_line'
 import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
-import { WarningMessage } from '../../common/warning_message'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
+import { WarningMessage } from '../common_sections/message_text/warning_message'
 
 const StyledButtonContainer = styled(ButtonContainer)`
   justify-content: space-between;
